fix lint error
